refactor: split metadata into v2 and v3 modules
