+++ src/zarr/core/metadata/__init__.py
@@ -1,0 +1,4 @@
+from .v2 import ArrayV2Metadata
+from .v3 import ArrayV3Metadata
+
+__all__ = ["ArrayV2Metadata", "ArrayV3Metadata"]
