+++ tests/v3/test_metadata/test_v3.py
@@ -7,6 +7,7 @@ from typing import TYPE_CHECKING, Literal
 from zarr.codecs.bytes import BytesCodec
 from zarr.core.buffer import default_buffer_prototype
 from zarr.core.chunk_key_encodings import DefaultChunkKeyEncoding, V2ChunkKeyEncoding
+from zarr.core.metadata.v3 import ArrayV3Metadata
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
@@ -18,9 +19,7 @@ if TYPE_CHECKING:
 import numpy as np
 import pytest
 
-from zarr.core.metadata import ArrayV3Metadata, parse_dimension_names
-from zarr.core.metadata import parse_fill_value_v3 as parse_fill_value
-from zarr.core.metadata import parse_zarr_format_v3 as parse_zarr_format
+from zarr.core.metadata.v3 import parse_dimension_names, parse_fill_value, parse_zarr_format
 
 bool_dtypes = ("bool",)
